restructure as python package
this will make it easier to add unit tests (and ultimately share
common code between commands)
