+++ setup.cfg
@@ -1,0 +1,9 @@
+[metadata]
+name = moc-sprint-tools
+version = attr:moc_sprint_tools.__version__
+
+[options.entry_points]
+console_scripts =
+	label-cards-in-sprint = moc_sprint_tools.label_cards_in_sprint:main
+	label-needs-description = moc_sprint_tools.label_needs_description:main
+	sort-cards-by-priority = moc_sprint_tools.sort_cards_by_priority:main
