+++ setup.py
@@ -1,0 +1,2 @@
+from setuptools import setup
+setup()
